Deprecate ElementFileParser trait in favor of loader and renderer
