+++ src/Element/Element.php
@@ -8,8 +8,8 @@
 
 namespace ActiveCollab\Shade\Element;
 
-use ActiveCollab\Shade\ElementFileParser;
 use ActiveCollab\Shade\Loader\LoaderInterface;
+use ActiveCollab\Shade\Loader\Result\LoaderResultInterface;
 use ActiveCollab\Shade\Project\ProjectInterface;
 use ActiveCollab\Shade\Renderer\RendererInterface;
 use ActiveCollab\Shade\Transformator\Transformator;
@@ -17,11 +17,14 @@ use ActiveCollab\Shade\Transformator\TransformatorInterface;
 
 abstract class Element implements ElementInterface
 {
-    use ElementFileParser;
-
     private $project;
     private $loader;
     private $renderer;
+
+    /**
+     * @var LoaderResultInterface|null
+     */
+    private $loadResult;
 
     public function __construct(
         ProjectInterface $project,
@@ -37,13 +40,44 @@ abstract class Element implements ElementInterface
         $this->path = $path;
 
         if ($load) {
-            $this->load();
+            $this->loadResult = $this->loader->load($this);
         }
     }
 
     public function renderBody(): string
     {
         return $this->renderer->renderElementBody($this);
+    }
+
+    /**
+     * @var string
+     */
+    private $index_file_path;
+
+    /**
+     * Get index file path.
+     *
+     * @return string
+     */
+    public function getIndexFilePath(): string
+    {
+        if (empty($this->index_file_path)) {
+            $this->index_file_path = is_dir($this->getPath()) ? $this->getPath() . '/index.md' : $this->getPath();
+        }
+
+        return $this->index_file_path;
+    }
+
+    public function isLoaded(): bool
+    {
+        return !empty($this->loadResult);
+    }
+
+    public function getProperty(string $name, string $default = null): ?string
+    {
+        return $this->loadResult instanceof LoaderResultInterface
+            ? $this->loadResult->getProperty($name, $default)
+            : $default;
     }
 
     public function getTransformator(): TransformatorInterface
