--- src/ElementFileParser.php
@@ -1,166 +1,0 @@
-<?php
-
-/*
- * This file is part of the Shade project.
- *
- * (c) A51 doo <info@activecollab.com>. All rights reserved.
- */
-
-namespace ActiveCollab\Shade;
-
-use ActiveCollab\Shade\Error\ElementFileNotFoundError;
-use ActiveCollab\Shade\Transformator\TransformatorInterface;
-
-trait ElementFileParser
-{
-    /**
-     * @var string
-     */
-    private $properties_separator = '================================================================';
-
-    /**
-     * Load indicator.
-     *
-     * @var bool
-     */
-    protected $is_loaded = false;
-
-    protected $properties = [];
-
-    /**
-     * Body text.
-     *
-     * @var string
-     */
-    protected $body;
-
-    public function getProperty(string $name, string $default = null): ?string
-    {
-        return $this->properties[$name] ?? $default;
-    }
-
-    /**
-     * Return element body.
-     *
-     * @return string
-     */
-    public function getBody()
-    {
-        return $this->body;
-    }
-
-    /**
-     * Return true if we loaded element's definition.
-     *
-     * @return bool
-     */
-    public function isLoaded()
-    {
-        return $this->is_loaded;
-    }
-
-    /**
-     * Load element's definition.
-     *
-     * @throws ElementFileNotFoundError
-     */
-    public function load()
-    {
-        if (empty($this->is_loaded)) {
-            $index_file = $this->getIndexFilePath();
-
-            if (is_file($index_file)) {
-                $this->body = file_get_contents($index_file);
-
-                $separator_pos = strpos($this->body, $this->properties_separator);
-
-                if ($separator_pos === false) {
-                    if (substr($this->body, 0, 1) == '*') {
-                        $properties_string = $this->body;
-                        $this->body = '';
-                    } else {
-                        $properties_string = '';
-                    }
-                } else {
-                    $properties_string = trim(substr($this->body, 0, $separator_pos));
-                    $this->body = trim(substr($this->body, $separator_pos + strlen($this->properties_separator)));
-                }
-
-                if ($properties_string) {
-                    $properties_lines = explode("\n", $properties_string);
-
-                    if (count($properties_lines)) {
-                        foreach ($properties_lines as $properties_line) {
-                            $properties_line = trim(trim($properties_line, '*')); // Clean up
-
-                            if ($properties_line) {
-                                $colon_pos = strpos($properties_line, ':');
-
-                                if ($colon_pos !== false) {
-                                    $this->loadProperty(trim(substr($properties_line, 0, $colon_pos)), trim(substr($properties_line, $colon_pos + 1)));
-                                }
-                            }
-                        }
-                    }
-                }
-
-                $this->body = trim($this->body);
-            } else {
-                throw new ElementFileNotFoundError($index_file);
-            }
-
-            $this->is_loaded = true;
-        }
-    }
-
-    /**
-     * Load property value.
-     *
-     * @param string $name
-     * @param string $value
-     */
-    private function loadProperty($name, $value)
-    {
-        $this->properties[$this->normalizePropertyName($name)] = $value;
-    }
-
-    private function normalizePropertyName(string $property_name): string
-    {
-        return Shade::underscore(str_replace(' ', '', $property_name));
-    }
-
-    /**
-     * @var string
-     */
-    private $index_file_path;
-
-    /**
-     * Get index file path.
-     *
-     * @return string
-     */
-    public function getIndexFilePath(): string
-    {
-        if (empty($this->index_file_path)) {
-            $this->index_file_path = is_dir($this->getPath()) ? $this->getPath() . '/index.md' : $this->getPath();
-        }
-
-        return $this->index_file_path;
-    }
-
-    // ---------------------------------------------------
-    //  Expectations
-    // ---------------------------------------------------
-
-    /**
-     * @return \ActiveCollab\Shade\Project\Project
-     */
-    abstract function &getProject();
-
-    /**
-     * @return string
-     */
-    abstract function getPath();
-
-    abstract function getTransformator(): TransformatorInterface;
-}
