+++ src/Loader/Result/LoaderResult.php
